% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipline.R
\name{calculate_all_lrlr}
\alias{calculate_all_lrlr}
\title{Calculate multi cell cell interactions of all lr-lr pairs}
\usage{
calculate_all_lrlr(
  Input,
  Coordinate,
  st_type = c("low", "high"),
  deconvolution_result = NULL,
  scRNA_mean_expression = NULL,
  ct1,
  ct2,
  ct3,
  relay_signal,
  lrpairs = NULL,
  n_permutations = 1000
)
}
\arguments{
\item{Input}{Normalized expression data matrix}

\item{Coordinate}{Normalized expression data matrix}

\item{st_type}{Normalized expression data matrix}
}
\value{
A dataframe of Receptor-Ligand signal weights of each cell or celltypes
}
\description{
Calculate multi cell cell interactions of all lr-lr pairs
}
\examples{
utils::data("test_exp_data", package = "stMCCC")
utils::data("test_anno", package = "stMCCC")
results = Calculate_Combined_RL_weight(test_exp_data[,1,drop = F], 'SingleCell')
results = Calculate_Combined_RL_weight(test_exp_data, test_anno$Label, 'CellType')
}
