% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipline.R
\name{Calculate_Combined_RL_weight}
\alias{Calculate_Combined_RL_weight}
\title{Calculate combined intracelluar Receptor-Liagnd siganal weight}
\usage{
Calculate_Combined_RL_weight(
  Input,
  cell_ids = NULL,
  PPI_db = NULL,
  Pro_loc_db = NULL,
  lr_db = NULL,
  f_edge = NULL,
  Kn = 2,
  cutoff1 = 0.5,
  cutoff2 = 0.9,
  max_steps = 5,
  max_num = 20,
  circuit_scale_factor = 0.1,
  adjust_alpha = TRUE,
  target_ratio = 0.5,
  parallel = TRUE,
  n_core = NULL
)
}
\arguments{
\item{Input}{A matrix containing gene expression data. (Recommend scRNA_mean_expression)}

\item{cell_ids}{The col.names selected to be calculate. If NULL, it will be selected all columns.}

\item{PPI_db}{A database of protein-protein interaction.
If NULL, it will be inferred from the data which contains a column called 'functional' which means stimulation, inhibition or unknown function by 1,-1,0.
If you want to provide yourself database, it should contain a 'functional' column and set the parameter 1.}

\item{Pro_loc_db}{A database of protein location. If NULL, it will be inferred from the data.
Default is organized Uniprot database which contains a column called 'location' which contains param 'Membrane', 'Intracellular' and 'Nuluar'.
If you want to provide yourself database, it should contain a 'location' column and set param.}

\item{lr_db}{A database of ligand-receptor. If NULL, it will be inferred from the data.
Default is the lr_db database from 'Spatalk' which contains two columns called 'ligand' and 'receptor'}

\item{f_edge}{A function to calculate score of each edge, default is Hill function.}

\item{Kn}{The parameter of Hill function. Default is 2.}

\item{cutoff1}{A quantile to select the probility of R-L pairs over cutoff. Default is 0.5.}

\item{cutoff2}{A quantile to build a subgraph containing genes over cutoff. Default is 0.9.}

\item{max_steps}{The max steps to be find in each R-L signal networks. Default is 5.}

\item{max_num}{The number of paths to be selected to calculate path scores. Default is 20.}

\item{circuit_scale_factor}{Adjustment of scaling factor for circuit score. Default is 0.1.}

\item{adjust_alpha}{A logical parameter that determines whether to adjust alpha. Default is TRUE.}

\item{target_ratio}{The goal of alpha adjustment. Default is 0.5.}

\item{parallel}{A logical parameter that determines whether to do parallel. Default is TRUE.}

\item{n_core}{The number of cores used for parallel computing. If NULL, it will be detectCores() - 2.}
}
\value{
A dataframe of Receptor-Ligand signal weights of each cell_id
}
\description{
Calculate combined intracelluar Receptor-Liagnd siganal weight
}
\examples{
utils::data("test_scRNA_mean_expression", package = "stMCCC")
results = Calculate_Combined_RL_weight(test_scRNA_mean_expression)
}
