% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipline.R
\name{Calculate_double_CCC}
\alias{Calculate_double_CCC}
\title{Calculate double cell cell communication}
\usage{
Calculate_double_CCC(
  ct1,
  ct2,
  spots_celltype_matrix,
  all_celltype_expression_list,
  double_spots_pairs,
  lrpairs,
  n_no_interact = 10000,
  p_cutoff = 0.95
)
}
\arguments{
\item{ct1}{Sender cell type name}

\item{ct2}{Receiver cell type name}

\item{spots_celltype_matrix}{Result from the function 'Reconstruct_RCTD'.}

\item{all_celltype_expression_list}{Result from the function 'Get_reconstruct_ST_expression_ct_list'}

\item{double_spots_pairs}{Result from the function 'Get_double_spots_pairs'}

\item{lrpairs}{Database of lrpairs}

\item{n_no_interact}{Background interaction spots pairs}

\item{p_cutoff}{A cutoff for zero-inflated gamma test}
}
\value{
A.
}
\description{
Calculate double cell cell communication
}
\examples{
all_celltype_expression_list = Calculate_double_CCC(st_exp_matrix,ct1,ct2, spots_celltype_matrix,all_celltype_expression_list, double_spots_pairs,lrpairs)
}
