% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipline.R
\name{calculate_multi_cci}
\alias{calculate_multi_cci}
\title{Calculate muti-cellular interactions in celltype resolution}
\usage{
calculate_multi_cci(
  Input,
  deconvolution_result,
  scRNA_mean_expression,
  relay_signal = NULL,
  lrpairs = NULL
)
}
\arguments{
\item{Input}{Normalized expression data matrix}

\item{lrpairs}{A database of ligand-receptor.
Default is the lrpairs database from 'Spatalk' which contains two columns called 'ligand' and 'receptor'}

\item{Anno}{If mode is 'CellType', you need to provide the annotation of cells/spots}

\item{PPI_database}{A database of protein-protein interaction.
Default is omnipath ppi database which contains a column called 'functional' which means stimulation, inhibition or unknown function by 1,-1,0.
If you want to provide yourself database, it should contain a 'functional' column and set the parameter 1.}

\item{Location_database}{A database of protein location.
Default is organized Uniprot database which contains a column called 'father_location' which contains three param 'Membrane', 'Intracellular' and 'Nuluar'.
If you want to provide yourself database, it should contain a 'father_location' column and set param.}

\item{f_edge}{A function to calculate score of each edge, default is Hill function.}

\item{expression_cutoff}{A quantile to build a subgraph containing genes over cutoff. Default is 0.9.}

\item{lr_expression_cutoff}{A quantile to select the probility of R-L pairs over cutoff. Default is 0.5.}

\item{max_steps}{find_simple_path max steps in each R-L pairs}

\item{max_num}{number of paths to calculate path scores}

\item{alpha}{Path score and circuit score coefficients}
}
\value{
A dataframe of Receptor-Ligand signal weights of each cell or celltypes
}
\description{
Calculate muti-cellular interactions in celltype resolution
}
\examples{
utils::data("test_exp_data", package = "stMCCC")
utils::data("test_anno", package = "stMCCC")
results = Calculate_Combined_RL_weight(test_exp_data[,1,drop = F], 'SingleCell')
results = Calculate_Combined_RL_weight(test_exp_data, test_anno$Label, 'CellType')
}
