% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipline.R
\name{calculate_single_ct_single_lrlr}
\alias{calculate_single_ct_single_lrlr}
\title{Calculate multi cell cell single cell interactions}
\usage{
calculate_single_ct_single_lrlr(
  Input,
  Coordinate,
  deconvolution_result,
  ct1,
  ct2,
  ct3,
  L1,
  R1,
  L2,
  R2,
  scRNA_mean_expression,
  relay_signal
)
}
\arguments{
\item{Input}{Normalized expression data matrix}
}
\value{
A dataframe of Receptor-Ligand signal weights of each cell or celltypes
}
\description{
Calculate multi cell cell single cell interactions
}
\examples{
utils::data("test_exp_data", package = "stMCCC")
utils::data("test_anno", package = "stMCCC")
results = Calculate_Combined_RL_weight(test_exp_data[,1,drop = F], 'SingleCell')
results = Calculate_Combined_RL_weight(test_exp_data, test_anno$Label, 'CellType')
}
