% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipline.R
\name{calculate_tri_cci_customized}
\alias{calculate_tri_cci_customized}
\title{Calculate Triple Cell-Cell Interaction (CCI) with Customized Parameters}
\usage{
calculate_tri_cci_customized(
  Input,
  Coordinate,
  st_type = c("low", "high"),
  relay_signal,
  n_permutations = 1000,
  future_globals_maxSize = 8e+09,
  n_neighbor = 6,
  parallel = FALSE,
  n_cores = NULL,
  Filter = TRUE,
  cell_type_list1 = NULL,
  cell_type_list2 = NULL,
  cell_type_list3 = NULL,
  lrpair_list1 = NULL,
  lrpair_list2 = NULL,
  deconvolution_result = NULL,
  scRNA_mean_expression = NULL,
  all_celltype_expression_list = NULL,
  double_lrpairs = NULL,
  verbose = TRUE
)
}
\arguments{
\item{Input}{A matrix containing gene expression data.(Recommend normalized data)}

\item{Coordinate}{A data frame containing the spatial coordinates of spots, with columns "x" and "y".}

\item{st_type}{A character string indicating the type of spatial transcriptomics data, either "low" or "high". Default is "low".}

\item{relay_signal}{A data frame containing relay signal information.}

\item{n_permutations}{An integer specifying the number of permutations for background score calculation. Default is 1000.}

\item{future_globals_maxSize}{A numeric value specifying the maximum size of global variables in future processing. Default is 8e9.}

\item{n_neighbor}{An integer specifying the number of neighbors to find. Default is 6.}

\item{parallel}{A logical value indicating whether to use parallel processing. Default is FALSE.}

\item{n_cores}{An integer specifying the number of cores to use in parallel processing. If NULL, it will use all available cores minus 1.}

\item{Filter}{A logical value indicating whether to filter the data. Default is TRUE.}

\item{cell_type_list1}{A character vector specifying the first list of cell types. If NULL, it will be inferred from the data.}

\item{cell_type_list2}{A character vector specifying the second list of cell types. If NULL, it will be inferred from the data.}

\item{cell_type_list3}{A character vector specifying the third list of cell types. If NULL, it will be inferred from the data.}

\item{lrpair_list1}{A data frame containing the first list of ligand-receptor pairs. If NULL, it will use the built-in data.}

\item{lrpair_list2}{A data frame containing the second list of ligand-receptor pairs. If NULL, it will use the built-in data.}

\item{deconvolution_result}{A matrix or data frame containing deconvolution results. Required if st_type is "low".}

\item{scRNA_mean_expression}{A matrix or data frame containing single-cell RNA-seq mean expression data. Required if st_type is "low".}

\item{all_celltype_expression_list}{A list containing the expression data of each cell type. If NULL, it will be calculated.}

\item{double_lrpairs}{A data frame containing double ligand-receptor pairs. If NULL, it will be generated.}

\item{verbose}{A logical value indicating whether to print progress messages. Default is TRUE.}
}
\value{
A data.table containing the calculated triple cell-cell interaction scores.
}
\description{
This function calculates the triple cell-cell interaction scores based on the provided input data.
It supports both parallel and sequential processing, and can handle different types of spatial transcriptomics data.
}
\examples{
result <- calculate_tri_cci_customized(Input = expr_data, Coordinate = coord_data, 
                                       relay_signal = relay_data, st_type = "low",
                                       deconvolution_result = deconv_data,
                                       scRNA_mean_expression = scRNA_data)

}
