% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipline.R
\name{Calculate_triple_CCC}
\alias{Calculate_triple_CCC}
\title{Calculate triple cell cell communication}
\usage{
Calculate_triple_CCC(
  ct1,
  ct2,
  ct3,
  all_celltype_expression_list,
  triple_interaction_pairs,
  lr_rl_lr_list
)
}
\arguments{
\item{ct1}{Sender cell type name}

\item{ct2}{Mid cell type name}

\item{ct3}{Receiver cell type name}

\item{spots_celltype_matrix}{Result from the function 'Reconstruct_RCTD'.}

\item{double_spots_pairs}{Result from the function 'Get_double_spots_pairs'}

\item{mid_spots_cut}{Database of lrpairs}

\item{bg_mid_spots_cut_min}{Background interaction spots pairs}

\item{p_cutoff}{A cutoff for zero-inflated gamma test}
}
\value{
A.
}
\description{
Calculate triple cell cell communication
}
\examples{
all_celltype_expression_list = Calculate_double_CCC(st_exp_matrix,ct1,ct2, spots_celltype_matrix,all_celltype_expression_list, double_spots_pairs,lrpairs)
}
