% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipline.R
\name{Construct_spots_celltype}
\alias{Construct_spots_celltype}
\title{Selected cell types in each spot.}
\usage{
Construct_spots_celltype(RCTD_result, max_n_types = 3, cell_ratio_cutoff = 0.1)
}
\arguments{
\item{RCTD_result}{Result from the function 'load_ccRCC_RCTD_result'}

\item{max_n_types}{The most cell types will be remained in a spot}

\item{cell_ratio_cutoff}{Cell ratio must higer than a cutoff. Default = 0.1.}
}
\value{
A matrix containing the cell type name in each spot.
}
\description{
Selected cell types in each spot.
}
\examples{
spots_celltype_matrix = Construct_spots_celltype(coords)
}
