% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{Get_double_spots_pairs}
\alias{Get_double_spots_pairs}
\title{Get interaction double spots pairs}
\usage{
Get_double_spots_pairs(coords, n_hop = 1)
}
\arguments{
\item{coords}{Coordinate of the Visium data}

\item{n_hop}{Selected the spots around a single spot. Default n_hop = 1. n_hop can be 1,2,3.}
}
\value{
A list
}
\description{
Get interaction double spots pairs
}
\examples{
double_spots_pairs = Get_double_spots_pairs(coords)
}
