% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{Get_interaction_cutoff}
\alias{Get_interaction_cutoff}
\title{Get interaction distance cutoff}
\usage{
Get_interaction_cutoff(coords, n_hop = 1)
}
\arguments{
\item{coords}{Coordinate of the Visium data}

\item{n_hop}{Selected the spots around a single spot. Default n_hop = 1. n_hop can be 1,2,3.}
}
\value{
A distance length for the interaction area and print the mean of interacted spots of a single spot
}
\description{
Get interaction distance cutoff
}
\examples{
RCTD_result = load_ccRCC_RCTD_result() 
}
