% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipline.R
\name{Get_reconstruct_ST_expression_ct_list}
\alias{Get_reconstruct_ST_expression_ct_list}
\title{Constucted a expression list of every cell types in each spot.}
\usage{
Get_reconstruct_ST_expression_ct_list(
  st_exp_matrix,
  sc_AverageExp,
  deconvolution_reconsitution
)
}
\arguments{
\item{st_exp_matrix}{Expression matrix of ST dataset.}

\item{sc_AverageExp}{Mean expression from scRNA-seq dataset.}

\item{RCTD_reconsitution}{Result from the function 'Reconstruct_RCTD' or a matrix like it.}
}
\value{
A matrix containing the cell type name in each spot.
}
\description{
Constucted a expression list of every cell types in each spot.
}
\examples{
all_celltype_expression_list = Get_reconstruct_ST_expression_ct_list(st_exp_matrix,sc_AverageExp,RCTD_reconsitution)
}
