% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{load_ccRCC_RCTD_result}
\alias{load_ccRCC_RCTD_result}
\title{Load RCTD deconvelution result}
\usage{
load_ccRCC_RCTD_result(
  pathway =
    "/Users/tangzongli_macmini/Desktop/2025 work/My_packages/stMCCC/test/ccRCC/RCTD/RCTD_matrix.csv"
)
}
\arguments{
\item{pathway}{Pathway of the result.}
}
\value{
A matrix with n_spots * c_cell_types
}
\description{
Load RCTD deconvelution result
}
\examples{
RCTD_result = load_ccRCC_RCTD_result() 
}
