% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{load_ccRCC_scRNA_seq_data}
\alias{load_ccRCC_scRNA_seq_data}
\title{Load ccRCC reference scRNA-seq dataset}
\usage{
load_ccRCC_scRNA_seq_data(
  pathway =
    "/Users/tangzongli_macmini/Desktop/2025 work/My_packages/stMCCC/test/ccRCC_scRNA"
)
}
\arguments{
\item{pathway}{Pathway of the dataset.
In this pathway, there is a annotation file and a 'GSE159115_RAW' fold containing
filtered_gene_bc_matrices_h5 files}
}
\value{
A scRNA-seq seurat object of ccRCC dataset
}
\description{
Load ccRCC reference scRNA-seq dataset
}
\examples{
sc_seu_obj = load_ccRCC_scRNA_seq_data()
}
