% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{load_ccRCC_Visium_data}
\alias{load_ccRCC_Visium_data}
\title{Load ccRCC Visium test dataset}
\usage{
load_ccRCC_Visium_data(
  pathway = "/Users/tangzongli_macmini/Desktop/2025 work/My_packages/stMCCC/test/ccRCC"
)
}
\arguments{
\item{pathway}{Pathway of the dataset.
In this pathway, there is a filtered_feature_bc_matrix.h5 dgCmatrix file, a TLS annotation file
and a spatial fold containing tissue_positions_list.csv and tissue_lowres_image.png}
}
\value{
A seurat object of the test dataset
}
\description{
Load ccRCC Visium test dataset
}
\examples{
st_seu_obj = load_ccRCC_Visium_data()  # Return a seurat Object
}
