% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_hierarchical_all_celltypes_interaction}
\alias{plot_hierarchical_all_celltypes_interaction}
\title{Plot hierarchical interaction of all celltypes}
\usage{
plot_hierarchical_all_celltypes_interaction(celltype_multi_cci, Mode = NULL)
}
\arguments{
\item{celltype_multi_cci}{Result from the function 'calculate_multi_cci'. We need the interaction scores and triple cell pairs.}

\item{Mode}{'Score' or 'Num'. Represent the mean score or the number of selected lr-lr pairs in all cell type pairs}
}
\value{
A ggplot object
}
\description{
Plot hierarchical interaction of all celltypes
}
\examples{
plot = plot_hierarchical_all_celltypes_interaction(celltype_multi_cci, Mode = 'Score')
plot = plot_hierarchical_all_celltypes_interaction(celltype_multi_cci, Mode = 'Num')
}
