% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_hierarchical_all_celltypes_lr_lr}
\alias{plot_hierarchical_all_celltypes_lr_lr}
\title{Plot hierarchical selected interaction of all celltypes}
\usage{
plot_hierarchical_all_celltypes_lr_lr(celltype_multi_cci, plot_row = NULL)
}
\arguments{
\item{celltype_multi_cci}{Result from the function 'calculate_multi_cci'. We need the interaction scores, double_lrpais and triple cell pairs.}

\item{plot_row}{The charactor name of lr-lr or the number of lr-lr in double_lrpais.}
}
\value{
A ggplot object
}
\description{
Plot hierarchical selected interaction of all celltypes
}
\examples{
plot_hierarchical_all_celltypes_lr_lr(celltype_multi_cci, plot_row = 1)
plot_hierarchical_all_celltypes_lr_lr(celltype_multi_cci, plot_row = 'SEMA3F-NRP1-SEMA3F-PLXNA3')
}
