% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_hierarchical_selected_celltypes_interaction}
\alias{plot_hierarchical_selected_celltypes_interaction}
\title{Plot hierarchical interaction of selected celltypes}
\usage{
plot_hierarchical_selected_celltypes_interaction(
  ct1s,
  ct2s,
  ct3s,
  celltype_multi_cci,
  Mode = NULL
)
}
\arguments{
\item{ct1s}{A list of selected sender cell types.}

\item{ct2s}{A list of selected relay cell types.}

\item{ct3s}{A list of selected receiver cell types.}

\item{celltype_multi_cci}{Result from the function 'calculate_multi_cci'. We need the interaction scores and triple cell pairs.}

\item{Mode}{'Score' or 'Num'. Represent the mean score or the number of selected lr-lr pairs in all cell type pairs}
}
\value{
A ggplot object
}
\description{
Plot hierarchical interaction of selected celltypes
}
\examples{
ct1s = c('Tumor', 'Bcell','Tcell')
ct2s = c('Macro','Tcell')
ct3s = c('Endo_PLVAP','Plasma')
plot_hierarchical_all_celltypes_interaction(celltype_multi_cci, Mode = 'Score')
plot_hierarchical_all_celltypes_interaction(celltype_multi_cci, Mode = 'Num')
}
