% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipline.R
\name{Prepare_calculate_triple_CCC}
\alias{Prepare_calculate_triple_CCC}
\title{Prepare a object to calculate triple cell cell communication}
\usage{
Prepare_calculate_triple_CCC(
  ct1,
  ct2,
  ct3,
  all_celltype_expression_list,
  triple_interaction_pairs,
  p_cutoff = 0.95
)
}
\arguments{
\item{ct1}{Sender cell type name}

\item{ct2}{Mid cell type name}

\item{ct3}{Receiver cell type name}

\item{all_celltype_expression_list}{Result from the function 'Get_reconstruct_ST_expression_ct_list'.}

\item{triple_interaction_pairs}{Result from the function 'Get_double_spots_pairs'}

\item{p_cutoff}{A cutoff for zero-inflated gamma test}

\item{lrpairs1}{Provide the preceding LR pairs. A data frame containing two columns with the name 'Ligand' and 'Receptor'}

\item{lrpairs2}{Provide the following LR pairs. A data frame containing two columns with the name 'Ligand' and 'Receptor'}

\item{rl_database}{Database of rl signal database}

\item{bg_mid_spots_cut_min}{Background interaction spots pairs}
}
\value{
A.
}
\description{
Prepare a object to calculate triple cell cell communication
}
\examples{
all_celltype_expression_list = Calculate_double_CCC(st_exp_matrix,ct1,ct2, spots_celltype_matrix,all_celltype_expression_list, double_spots_pairs,lrpairs)
}
