% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipline.R
\name{Prepare_lr_rl_lr}
\alias{Prepare_lr_rl_lr}
\title{Prepare intracelluar Receptor Liagnd siganal database}
\usage{
Prepare_lr_rl_lr(
  lrpairs1,
  lrpairs2,
  rl_database,
  min_network_gene_num = 35,
  min_cutoff = 5
)
}
\arguments{
\item{lrpairs1}{Provide the preceding LR pairs. A data frame containing two columns with the name 'Ligand' and 'Receptor'}

\item{lrpairs2}{Provide the following LR pairs. A data frame containing two columns with the name 'Ligand' and 'Receptor'}

\item{rl_database}{Database of rl signal database}

\item{min_network_gene_num}{Background interaction spots pairs}
}
\value{
A.
}
\description{
Prepare intracelluar Receptor Liagnd siganal database
}
\examples{
all_celltype_expression_list = Calculate_double_CCC(st_exp_matrix,ct1,ct2, spots_celltype_matrix,all_celltype_expression_list, double_spots_pairs,lrpairs)
}
