% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipline.R
\name{Reconstruct_RCTD}
\alias{Reconstruct_RCTD}
\title{Reconstructed RCTD result based on 'Construct_spots_celltype'.}
\usage{
Reconstruct_RCTD(RCTD_result, spots_celltype_matrix)
}
\arguments{
\item{RCTD_result}{Result from the function 'load_ccRCC_RCTD_result'.}

\item{spots_celltype_matrix}{Result from the function 'Construct_spots_celltype'.}

\item{cell_ratio_cutoff}{Cell ratio must higer than a cutoff. Default = 0.1.}
}
\value{
A matrix containing the cell type name in each spot.
}
\description{
Reconstructed RCTD result based on 'Construct_spots_celltype'.
}
\examples{
spots_celltype_matrix = Reconstruct_RCTD(coords)
}
