% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipline.R
\name{validate_inputs}
\alias{validate_inputs}
\title{Validate input data for calculate_all_lrlr function.}
\usage{
validate_inputs(
  Input,
  Coordinate,
  st_type,
  deconvolution_result = NULL,
  scRNA_mean_expression = NULL,
  lr_db = NULL
)
}
\arguments{
\item{Input}{A matrix or data frame containing gene expression data(Recommend normalized data).}

\item{Coordinate}{Coordinate matrix.}

\item{st_type}{A character string indicating the type of spatial transcriptomics data, either "low" or "high". Default is "low".}

\item{deconvolution_result}{A matrix or data frame containing deconvolution results. Required if st_type is "low".}

\item{scRNA_mean_expression}{A matrix or data frame containing single-cell RNA-seq mean expression data. Required if st_type is "low".}

\item{lr_db}{A data frame containing ligand-receptor pairs.}
}
\value{
TRUE if all inputs are valid, otherwise stop with an error message.
}
\description{
Validate input data for calculate_all_lrlr function.
}
