% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipline.R
\name{Calculate_Combined_RL_weight}
\alias{Calculate_Combined_RL_weight}
\title{Calculate combined intracelluar Receptor-Liagnd siganal weight}
\usage{
Calculate_Combined_RL_weight(
  Input,
  cell_ids = NULL,
  PPI_db = NULL,
  Pro_loc_db = NULL,
  lr_db = NULL,
  f_edge = NULL,
  Kn = 2,
  cutoff1 = 0.5,
  cutoff2 = 0.9,
  max_steps = 5,
  max_num = 20,
  circuit_scale_factor = 0.1,
  adjust_alpha = TRUE,
  target_ratio = 0.5,
  parallel = TRUE,
  n_core = NULL
)
}
\arguments{
\item{Input}{Normalized expression data matrix}

\item{Pro_loc_db}{A database of protein location.
Default is organized Uniprot database which contains a column called 'father_location' which contains three param 'Membrane', 'Intracellular' and 'Nuluar'.
If you want to provide yourself database, it should contain a 'father_location' column and set param.}

\item{lr_db}{A database of ligand-receptor.
Default is the lr_db database from 'Spatalk' which contains two columns called 'ligand' and 'receptor'}

\item{f_edge}{A function to calculate score of each edge, default is Hill function.}

\item{cutoff1}{A quantile to select the probility of R-L pairs over cutoff. Default is 0.5.}

\item{max_steps}{find_simple_path max steps in each R-L pairs}

\item{max_num}{number of paths to calculate path scores}

\item{Anno}{If mode is 'CellType', you need to provide the annotation of cells/spots}

\item{PPI_database}{A database of protein-protein interaction.
Default is omnipath ppi database which contains a column called 'functional' which means stimulation, inhibition or unknown function by 1,-1,0.
If you want to provide yourself database, it should contain a 'functional' column and set the parameter 1.}

\item{expression_cutoff}{A quantile to build a subgraph containing genes over cutoff. Default is 0.9.}

\item{alpha}{Path score and circuit score coefficients}
}
\value{
A dataframe of Receptor-Ligand signal weights of each cell or celltypes
}
\description{
Calculate combined intracelluar Receptor-Liagnd siganal weight
}
\examples{
utils::data("test_exp_data", package = "stMCCC")
utils::data("test_anno", package = "stMCCC")
results = Calculate_Combined_RL_weight(test_exp_data[,1,drop = F], 'SingleCell')
results = Calculate_Combined_RL_weight(test_exp_data, test_anno$Label, 'CellType')
}
