% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipline.R
\name{validate_inputs}
\alias{validate_inputs}
\title{Validate input data for calculate_all_lrlr function.}
\usage{
validate_inputs(
  Input,
  Coordinate,
  st_type,
  deconvolution_result = NULL,
  scRNA_mean_expression = NULL
)
}
\arguments{
\item{Input}{Normalized expression data matrix.}

\item{Coordinate}{Coordinate matrix.}

\item{st_type}{Spatial transcriptome data type ('low' or 'high').}

\item{deconvolution_result}{Deconvolution result matrix.}

\item{scRNA_mean_expression}{Mean expression matrix from scRNA-seq.}
}
\value{
TRUE if all inputs are valid, otherwise stop with an error message.
}
\description{
Validate input data for calculate_all_lrlr function.
}
